pyproject.toml
[tool.poetry]
name = "fedcot-align"
version = "0.1.0"
description = "Federated Alignment of Chain-of-Thought Reasoning for Hallucination Reduction"
authors = ["Senior Research Engineer <research@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = ">=2.4.0"
transformers = ">=4.44.0"
flwr = { extras = ["simulation"], version = ">=1.10.0" }
datasets = ">=3.0.0"
pyyaml = "^6.0.1"
wandb = ">=0.17.0"
scipy = "^1.13.0"
numpy = "^1.26.0"
scikit-learn = "^1.5.0"
sentencepiece = "^0.2.0"
evaluate = "^0.4.2"
rouge_score = "^0.1.2"
accelerate = "^0.34.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.0.0"
isort = "^5.13.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"